development-ml/create-ml-development-db-tables-app/requirements.txt
python-dotenv==1.0.1
SQLAlchemy==2.0.37
asyncpg==0.30.0
pydantic-settings==2.7.1
bcrypt==4.2.1
sqlalchemy-utils==0.41.2
cryptography==44.0.2